+++ mmcls/datasets/samplers/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .builder import SAMPLERS
 from .distributed_sampler import DistributedSampler
 from .repeat_aug import RepeatAugSampler
 
-__all__ = ('DistributedSampler', 'RepeatAugSampler', 'SAMPLERS')
+__all__ = ('DistributedSampler', 'RepeatAugSampler')
