+++ mmcls/datasets/samplers/distributed_sampler.py
@@ -2,7 +2,7 @@
 import torch
 from torch.utils.data import DistributedSampler as _DistributedSampler
 
-from mmcls.datasets.samplers import SAMPLERS
+from mmcls.datasets import SAMPLERS
 
 
 @SAMPLERS.register_module()
